backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
gunicorn==21.2.0